Align gfx942 split-KV tests with the aiter op-test format
Match the shared fake p dtype to the real op, reject non-finite bench
comparisons, and sweep the kernel with a torch reference plus markdown tables.

+++ aiter/ops/mha.py
@@ -1111,7 +1111,7 @@ def gen_fmha_v3_varlen_fwd_fake_tensor(
         p_shape = (num_heads, total_q, max_seqlen_k)
         p_tensor = torch.empty(p_shape, device=device, dtype=torch.uint8)
     else:
-        p_tensor = torch.empty((0,), device=device)
+        p_tensor = torch.empty((0,), device=device, dtype=dtype)
 
     rng_state_tensor = torch.empty((2,), device=device, dtype=torch.int64)
